wip(app): content tags, profile options, and a priced-photo test
Parked on staging, not reviewed or verified by me. The picco-v2 pointer moves
to that submodule's own staging branch.

+++ mobile/lib/features/discovery/presentation/pages/event_pictures_page.dart
@@ -41,7 +41,7 @@ List<Photo> photosOfEvent(EventDiscovery event) => [
           p.imageId,
           p.url,
           event.photographerId,
-          p.price.toDouble(),
+          p.price,
           '',
           null,
           // Everything reachable from the public feed is public; the badge over
